fix(security): replace cleartext HTTP with raw sockets, harden deferred items
Eliminate global cleartextTrafficPermitted="true" by refactoring
NetworkScanner and ConnectionProber to use raw java.net.Socket for
local HTTP probes (bypasses Android's best-effort cleartext policy).
Also resolves 6 deferred review items: XML tool name validation,
response size cap, fallbackToDestructiveMigration removal,
SessionStateGuard expect-to-Result, provider_backoff_ms Int-to-Long,
and minSdk bump to API 28.

- Add readCapped(), rawHttpGet(), decodeChunked() to NetworkScanner
- Refactor ConnectionProber with raw socket HEAD for HTTP URLs
- Lock down network_security_config.xml (cleartextTrafficPermitted=false)
- Add NEARBY_WIFI_DEVICES permission for Android 16+ prep
- Filter unrecognised XML tool calls in session.rs (prompt injection fix)
- Fix SessionStateGuard::take() partial extraction with combined match
- Change provider_backoff_ms from Int to Long across 14 files
- Remove fallbackToDestructiveMigration from ZeroClawDatabase
- Bump minSdk 26 -> 28 in app/ and lib/ modules
- 17 new NetworkScanner unit tests, 358 Rust tests passing

+++ app/src/main/java/com/zeroclaw/android/data/repository/SettingsRepository.kt
@@ -839,7 +839,7 @@ interface SettingsRepository {
      *
      * @param ms Backoff duration in milliseconds.
      */
-    suspend fun setReliabilityBackoffMs(ms: Int)
+    suspend fun setReliabilityBackoffMs(ms: Long)
 
     /**
      * Updates the reliability API keys JSON.
