+++ app/src/test/java/com/zeroclaw/android/ui/screen/onboarding/OnboardingViewModelProviderTest.kt
@@ -493,7 +493,7 @@ private class TestSettingsRepository : SettingsRepository {
 
     override suspend fun setProxyServiceSelectors(selectors: String) { /* no-op */ }
 
-    override suspend fun setReliabilityBackoffMs(ms: Int) { /* no-op */ }
+    override suspend fun setReliabilityBackoffMs(ms: Long) { /* no-op */ }
 
     override suspend fun setReliabilityApiKeysJson(json: String) { /* no-op */ }
